% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{alpha_b}
\alias{alpha_b}
\title{Return the conditional density of a normal random variable, given that its
value is between a and b, evaluated at z_b.}
\usage{
alpha_b(mean, sd, a, b)
}
\arguments{
\item{mean}{Mean of normal distribution.}

\item{sd}{Standard deviation of normal distribution.}

\item{a}{Unstandardized lower truncation limit.}

\item{b}{Unstandardized upper truncation limit.}
}
\description{
Return the conditional density of a normal random variable, given that its
value is between a and b, evaluated at z_b.
}
\examples{
alpha_b(mean = 1, sd = 1, a = -5, b = 5)
}
