% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alphah}
\alias{alphah}
\title{Return the conditional density of a normal random variable, given that its value is between ul and uh, evaluated at ush.}
\usage{
alphah(.mean, .sd, ul, uh)
}
\arguments{
\item{.mean}{Mean of normal distribution.}

\item{.sd}{Standard deviation of normal distribution.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}
}
\description{
Return the conditional density of a normal random variable, given that its value is between ul and uh, evaluated at ush.
}
