% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alphal}
\alias{alphal}
\title{Return the conditional density of a normal random variable, given that its value is between ul and uh, evaluated at usl.}
\usage{
alphal(m, s, ul, uh)
}
\arguments{
\item{m}{Mean of normal distribution.}

\item{s}{Standard deviation of normal distribution.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}
}
\description{
Return the conditional density of a normal random variable, given that its value is between ul and uh, evaluated at usl.
}
