% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Applies the bias correction, K^{-1}AK, of Godwin et al. and
calculates bias of the bias-corrected MLE.}
\usage{
calculate_bias(.mean, .sd, ul, uh, n, sim.reps, type)
}
\arguments{
\item{.mean}{Mean of underlying normal distribution.}

\item{.sd}{Standard deviation of underlying normal distribution.}

\item{ul}{Lower truncation limit.}

\item{uh}{Upper truncation limit.}

\item{n}{Number of observations.}

\item{sim.reps}{Number of repetitions to estimate bias of the bias-corrected MLE.}

\item{type}{"meanub", "stddevub" for bias of Cordiero estimator for mean or standard deviation respectively.
"meanb", "stddevb" for bias of MLE of mean and standard deviation.}
}
\description{
Applies the bias correction, K^{-1}AK, of Godwin et al. and
calculates bias of the bias-corrected MLE.
}
\references{
Ryan T. Godwin (2016) Bias reduction for the maximum likelihood estimator of the doubly-truncated Poisson distribution,
Communications in Statistics - Theory and Methods, 45:7, 1887-1901, DOI: 10.1080/03610926.2013.867999
}
