% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cordeiro_bias}
\alias{cordeiro_bias}
\title{Calculates the Cordiero bias-correction for the MLE of the mean and standard deviation of a
doubly-truncated normal.}
\usage{
cordeiro_bias(.mean, .sd, .ul, .uh, .n)
}
\arguments{
\item{.mean}{Mean of underlying normal distribution.}

\item{.sd}{Standard deviation of underlying normal distribution.}

\item{.ul}{Lower truncation limit.}

\item{.uh}{Upper truncation limit.}

\item{.n}{Number of observations.}
}
\description{
Calculates the Cordiero bias-correction for the MLE of the mean and standard deviation of a
doubly-truncated normal.
}
