% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cumulants_check}
\alias{cumulants_check}
\title{Ensures that the cumulants calculated in the Deriv package match those calculated by hand.
PUT IN TEST_MOMENTS.R}
\usage{
cumulants_check(m, s, ul, uh, n)
}
\arguments{
\item{m}{Mean of underlying normal distribution.}

\item{s}{Standard deviation of underlying normal distribution.}

\item{ul}{Lower truncation limit.}

\item{uh}{Upper truncation limit.}

\item{n}{Number of observations.}
}
\description{
Ensures that the cumulants calculated in the Deriv package match those calculated by hand.
PUT IN TEST_MOMENTS.R
}
