% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_jeffreys_mcmc}
\alias{estimate_jeffreys_mcmc}
\title{Finds the MAP estimates for mu and sigma of the full normal distribution, given data from a truncated normal.}
\usage{
estimate_jeffreys_mcmc(x, mean.start, sd.start, ci.left, ci.right, ...)
}
\arguments{
\item{x}{Vector of observations from truncated normal}

\item{mean.start}{Initial value for mu.}

\item{sd.start}{Initial value for sigma.
MM: Let's format ci.left and ci.right as a single numerical arg, ci.level (default 0.95), and assume they want symmetric}

\item{ci.left}{Number between 0 and 1 . Left end of a confidence interval for each parameter estimate.}

\item{ci.right}{Number between 01 and 1. Right end of a confidence interval for each parameter estimate. By default 1-ci.left.}

\item{...}{Parameters to pass to sampling()}
}
\description{
Finds the MAP estimates for mu and sigma of the full normal distribution, given data from a truncated normal.
}
\references{
https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html
}
