% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_A}
\alias{find_A}
\title{Finds the A matrix from Godwin et al.}
\usage{
find_A(.mean, .sd, ul, uh, n)
}
\arguments{
\item{.mean}{Mean of underlying normal distribution.}

\item{.sd}{Standard deviation of underlying normal distribution.}

\item{ul}{Lower truncation limit.}

\item{uh}{Upper truncation limit.}

\item{n}{Number of observations.}
}
\description{
Finds the A matrix from Godwin et al.
}
\references{
Ryan T. Godwin (2016) Bias reduction for the maximum likelihood estimator of the doubly-truncated Poisson distribution,
Communications in Statistics - Theory and Methods, 45:7, 1887-1901, DOI: 10.1080/03610926.2013.867999
}
