% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_K}
\alias{find_K}
\title{Finds the Fisher Information matrix on n observations from a truncated normal distribution.}
\usage{
find_K(m, s, ul, uh, n)
}
\arguments{
\item{m}{Mean of underlying normal distribution.}

\item{s}{Standard deviation of underlying normal distribution.}

\item{ul}{Lower truncation limit.}

\item{uh}{Upper truncation limit.}

\item{n}{Number of observations.}
}
\description{
Finds the Fisher Information matrix on n observations from a truncated normal distribution.
}
