% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cumulants}
\alias{get_cumulants}
\title{Finds cumulants of the full log-likelihood.}
\usage{
get_cumulants(d_str, m, s, ul, uh, n)
}
\arguments{
\item{d_str}{Type of derivative to take. Consists of characters "m" and "s"}

\item{m}{Mean of underlying normal distribution.}

\item{s}{Standard deviation of underlying normal distributin.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}
}
\description{
Finds cumulants of the full log-likelihood.
}
