% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_deriv}
\alias{get_deriv}
\title{Finds the derivative of the factor of log-likelihood independent of data.}
\usage{
get_deriv(d_str)
}
\arguments{
\item{dstr}{Type of derivative to take. Consists of characters "m" and "s"}
}
\description{
Finds the derivative of the factor of log-likelihood independent of data.
}
