% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ll}
\alias{ll}
\title{Calculates factor of log-likelihood independent of the data.
We assume that the data are distributed according to a normal distribution of mean m, standard
deviation s, truncated at lower and upper limits, ul and uh, respectively.}
\usage{
ll(m, s, ul, uh, n)
}
\arguments{
\item{m}{Mean of normal distribution.}

\item{s}{Standard deviation of normal distribution.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}

\item{n}{Number of observations.}
}
\description{
Calculates factor of log-likelihood independent of the data.
We assume that the data are distributed according to a normal distribution of mean m, standard
deviation s, truncated at lower and upper limits, ul and uh, respectively.
}
