% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{neg_log_post}
\alias{neg_log_post}
\title{Returns the value of the negative log-posterior with the Jeffrey's prior.}
\usage{
neg_log_post(.pars, .x, .a, .b)
}
\arguments{
\item{.pars}{Vector specifying mu and sigma.}

\item{.x}{Data to use log likelihood calculation.}

\item{.a}{Left truncation limit.}

\item{.b}{Right truncation limit.}
}
\description{
Returns the value of the negative log-posterior with the Jeffrey's prior.
}
