% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nlpost_Jeffreys}
\alias{nlpost_Jeffreys}
\title{Auxillary function to calculate the negative log posterior with Jeffrey's prior.}
\usage{
nlpost_Jeffreys(.pars, .x, .a, .b)
}
\arguments{
\item{.pars}{Vector of parameters specifying mean and standard deviation.}

\item{.x}{Data to use log likelihood calculation.}

\item{.a}{Left truncation limit.}

\item{.b}{Right truncation limit.}
}
\description{
Auxillary function to calculate the negative log posterior with Jeffrey's prior.
}
