% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{samp_trunc}
\alias{samp_trunc}
\title{Takes samples from a normal distribution of mean m and standard deviation s.
Rejects samples greater than uh or less than ul.}
\usage{
samp_trunc(nsamp, mu, sigma, ul, uh)
}
\arguments{
\item{nsamp}{Number of observations to draw from normal distribution.}

\item{mu}{Mean of normal distribution}

\item{sigma}{Standard deviation of normal distribution.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}
}
\description{
Takes samples from a normal distribution of mean m and standard deviation s.
Rejects samples greater than uh or less than ul.
}
