% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trunc_est}
\alias{trunc_est}
\title{Estimates the posterior modes for the parameters of the underlying normal
distribution, given truncated data}
\usage{
trunc_est(x, mean_start = 0, sd_start = 1, ci_level = 0.95, a, b, ...)
}
\arguments{
\item{x}{Vector of observations from truncated normal}

\item{mean_start}{Initial value for mu.}

\item{sd_start}{Initial value for sigma.}

\item{ci_level}{Number between 0.5 and 1. Gives a 100(ci_level)\% confidence
interval.}

\item{a}{Left truncation limit.}

\item{b}{Right truncation limit.}

\item{...}{Parameters to pass to sampling()}
}
\description{
Estimates the posterior modes for the parameters of the underlying normal
distribution, given truncated data
}
\examples{
x <- truncnorm::rtruncnorm(100, a = 0, b = 2, mean = 0.5, sd = 0.5)
trunc_est(x, a = 0, b = 2)

}
\references{
https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html
}
