% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncNormalMoments-package.R
\docType{package}
\name{truncNormalMoments-package}
\alias{truncNormalMoments}
\alias{truncNormalMoments-package}
\title{truncNormalMoments: Estimates mean and standard deviation for a truncated normal distribution using Stan}
\description{
This package finds the MAP parameter estimates for the mean and standard deviation for a normal distribution, given data from the truncated distribution. The method used extends Bayesian methods for parameter estimation for a singly truncated normal distribution [see Zhou, X., Giacometti, R., Fabozzi, F. J., & Tucker, A. H. (2014). Bayesian estimation of truncated data with applications to operational risk measurement. Quantitative Finance, 14(5), 863-888. \doi{10.1080/14697688.2012.752103}]. We extend it by considering the doubly truncated normal distribution. We use the Jeffreys prior and find MAP estimates of the mean and standard deviation of a doubly truncated normal, with left and right truncations specified beforehand.
}
\references{
\insertRef{zhou2014}{truncNormalMoments}

\insertRef{stan2022}{truncNormalMoments}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonkt/truncNormalMoments}
  \item Report bugs at \url{https://github.com/leonkt/truncNormalMoments/issues}
}

}
\author{
\strong{Maintainer}: Leon Tran \email{leontk@stanford.edu}

Authors:
\itemize{
  \item Mika Braginsky \email{mika.br@gmail.com}
}

Other contributors:
\itemize{
  \item Maya Mathur \email{mmathur@stanford.edu} [contributor]
  \item Blakeley McShane \email{b-mcshane@kellogg.northwestern.edu} [contributor]
}

}
\keyword{internal}
