% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{usl}
\alias{usl}
\title{Returns the standardized lower truncation limit for a normal distribution.}
\usage{
usl(.mean, .sd, ul, uh)
}
\arguments{
\item{.mean}{Mean of normal distribution.}

\item{.sd}{Standard deviation of normal distribution.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}
}
\description{
Returns the standardized lower truncation limit for a normal distribution.
}
