% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Zb}
\alias{Zb}
\title{Returns the standardized upper truncation limit for a normal distribution.}
\usage{
Zb(.mean, .sd, .a, .b)
}
\arguments{
\item{.mean}{Mean of normal distribution.}

\item{.sd}{Standard deviation of normal distribution.}

\item{.a}{Unstandardized lower truncation limit.}

\item{.b}{Unstandardized upper truncation limit.}
}
\description{
Returns the standardized upper truncation limit for a normal distribution.
}
