% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{e_fisher}
\alias{e_fisher}
\title{Finds the Fisher information matrix contained in n samples from a truncated normal distribution.}
\usage{
e_fisher(mean, sd, n, a, b)
}
\arguments{
\item{mean}{Mean of underlying normal distribution.}

\item{sd}{Standard deviation of underlying normal distribution.}

\item{n}{Number of observations.}

\item{a}{Lower truncation limit.}

\item{b}{Upper truncation limit.}
}
\description{
Finds the Fisher information matrix contained in n samples from a truncated normal distribution.
}
\examples{
e_fisher(mean = 1, sd = 1, n = 10, a = -5, b = 5)
}
