% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cumulants}
\alias{get_cumulants}
\title{Finds cumulants of the full log-likelihood.}
\usage{
get_cumulants(d_str, .mean, .sd, ul, uh, n)
}
\arguments{
\item{d_str}{Type of derivative to take. Consists of characters "m" and "s"}

\item{.mean}{Mean of underlying normal distribution.}

\item{.sd}{Standard deviation of underlying normal distributin.}

\item{ul}{Unstandardized lower truncation limit.}

\item{uh}{Unstandardized upper truncation limit.}

\item{n}{Number of observations.}
}
\description{
Finds cumulants of the full log-likelihood.
}
