% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cumulants_deriv}
\alias{get_cumulants_deriv}
\title{Get derivatives w.r.t var of the cumulants.
Evaluates derivatives of cumulants at specific value.}
\usage{
get_cumulants_deriv(d_str, var, m, s, ul, uh, n)
}
\arguments{
\item{d_str}{Cumulant to take the derivative of}

\item{var}{Derivative of cumulant with respect to var}
}
\description{
Get derivatives w.r.t var of the cumulants.
Evaluates derivatives of cumulants at specific value.
}
