% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nlpost_Jeffreys}
\alias{nlpost_Jeffreys}
\title{Auxillary function to calculate the negative log posterior with Jeffrey's prior.}
\usage{
nlpost_Jeffreys(.pars, par2is = "sd", .x, .a, .b)
}
\arguments{
\item{.pars}{Vector of parameters specifying mu and sigma.}

\item{par2is}{"sd" if the second entry in .pars is the standard deviation, and "var" otherwise.}

\item{.x}{Data to use log likelihood calculation.}

\item{.a}{Left truncation limit.}

\item{.b}{Right truncation limit.}
}
\description{
Auxillary function to calculate the negative log posterior with Jeffrey's prior.
}
