% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nlpost_jeffreys}
\alias{nlpost_jeffreys}
\title{Calculate the negative log posterior under the Jeffreys prior}
\usage{
nlpost_jeffreys(mean, sd, x, a, b)
}
\arguments{
\item{mean}{Mean of normal distribution.}

\item{sd}{Standard deviation of normal distribution.}

\item{x}{Data to use log likelihood calculation.}

\item{a}{Left truncation limit.}

\item{b}{Right truncation limit.}
}
\description{
Calculate the negative log posterior under the Jeffreys prior
}
