% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prior}
\alias{prior}
\title{Returns the value of the Jeffreys prior.}
\usage{
prior(.pars, .x, .a, .b)
}
\arguments{
\item{.pars}{Vector of parameters specifying mean and standard deviation.}

\item{.x}{Data to use log likelihood calculation.}

\item{.a}{Left truncation limit.}

\item{.b}{Right truncation limit.}
}
\description{
Returns the value of the Jeffreys prior.
}
