% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prior_jeffreys}
\alias{prior_jeffreys}
\title{Returns the value of the Jeffreys prior.}
\usage{
prior_jeffreys(mean, sd, x, a, b)
}
\arguments{
\item{mean}{Mean of untruncated normal distribution}

\item{sd}{Standard deviation of untruncated normal distribution.}

\item{x}{Data to use log likelihood calculation.}

\item{a}{Left truncation limit.}

\item{b}{Right truncation limit.}
}
\description{
Returns the value of the Jeffreys prior.
}
