% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trunc_est}
\alias{trunc_est}
\title{Estimates the posterior modes for the parameters of the underlying normal distribution, given truncated data}
\usage{
trunc_est(x, mean.start = 0, sd.start = 1, ci.level = 0.95, a, b, ...)
}
\arguments{
\item{x}{Vector of observations from truncated normal}

\item{mean.start}{Initial value for mu.}

\item{sd.start}{Initial value for sigma.}

\item{ci.level}{Number between 0.5 and 1. Gives a 100(ci.level)\% confidence interval.}

\item{a}{Left truncation limit.}

\item{b}{Right truncation limit.}

\item{...}{Parameters to pass to sampling()}
}
\description{
Estimates the posterior modes for the parameters of the underlying normal distribution, given truncated data
}
\references{
https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html
}
