% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncnormbayes-package.R
\docType{package}
\name{truncnormbayes-package}
\alias{truncnormbayes}
\alias{truncnormbayes-package}
\title{truncnormbayes: Estimates Moments for a Truncated Normal Distribution using 'Stan'}
\description{
Finds the posterior modes for the mean and standard deviation for a truncated normal distribution with one or two known truncation points. The method used extends Bayesian methods for parameter estimation for a singly truncated normal distribution under the Jeffreys prior (see Zhou X, Giacometti R, Fabozzi FJ, Tucker AH (2014). "Bayesian estimation of truncated data with applications to operational risk measurement". \doi{10.1080/14697688.2012.752103}). This package additionally allows for a doubly truncated normal distribution.
}
\references{
\insertRef{zhou2014}{truncnormbayes}

\insertRef{stan2022}{truncnormbayes}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonkt/truncnormbayes}
  \item Report bugs at \url{https://github.com/leonkt/truncnormbayes/issues}
}

}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Leon Tran \email{leontk@stanford.edu}
}

Other contributors:
\itemize{
  \item Maya Mathur \email{mmathur@stanford.edu} [contributor]
}

}
\keyword{internal}
