% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ush}
\alias{ush}
\title{Returns the standardized upper truncation limit for a normal distribution.}
\usage{
ush(m, s, ul, uh)
}
\arguments{
\item{m}{Mean of normal distribution.}

\item{s}{Standard deviation of normal distribution.}

\item{uh}{Unstandardized upper truncation limit.}
}
\description{
Returns the standardized upper truncation limit for a normal distribution.
}
\examples{
Returns the z-score of the value 1, for a normal distribution of mean 0, standard deviation 2.
ush(0, 2, 1)
}
