% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{usl}
\alias{usl}
\title{Returns the standardized lower truncation limit for a normal distribution.}
\usage{
usl(m, s, ul, uh)
}
\arguments{
\item{m}{Mean of normal distribution.}

\item{s}{Standard deviation of normal distribution.}

\item{ul}{Unstandardized lower truncation limit.}
}
\description{
Returns the standardized lower truncation limit for a normal distribution.
}
